requirements.txt
aiodns==3.2.0
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
APScheduler==3.10.4
arrow==1.3.0
attrs==24.2.0
bravado==10.4.1
bravado-core==6.1.1
Bybit==0.2.12
ccxt==4.4.25
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
fqdn==1.5.1
frozenlist==1.5.0
idna==3.10
importlib_resources==6.4.5
isoduration==20.11.0
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
magic-filter==1.0.12
monotonic==1.6
msgpack==1.1.0
msgpack-python==0.5.6
multidict==6.1.0
propcache==0.2.0
pyaes==1.6.1
pybit==5.8.0
pycares==4.4.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
Pyrogram==2.0.106
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
setuptools==75.3.0
simplejson==3.19.3
six==1.16.0
swagger-spec-validator==3.0.4
TgCrypto==1.2.5
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzlocal==5.2
uri-template==1.3.0
urllib3==2.2.3
webcolors==24.8.0
websocket-client==1.8.0
yarl==1.17.0
